requirements.txt
python-dotenv>=1.0.1
pypdf>=5.2.0
cryptography>=3.1
langchain>=0.3.20
langchain-community>=0.3.19
langchain-text-splitters>=0.3.6
langchain-openai>=0.3.7
langchain-ollama>=0.2.3
langchain-huggingface>=0.1.2
faiss-cpu>=1.10.0
sentence-transformers>=3.4.1
arxiv>=2.1.3
streamlit>=1.42.0
pandas>=2.2.3
rouge-score>=0.1.2
jieba>=0.42.1
tqdm>=4.67.1
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
